+++ oneibl/one.py
@@ -1,3 +1,4 @@
+import abc
 import concurrent.futures
 import logging
 import os
@@ -5,15 +6,18 @@ from pathlib import Path, PurePath
 
 import requests
 import tqdm
+import pandas as pd
+import numpy as np
 
 import oneibl.params
 import oneibl.webclient as wc
 from alf.io import (AlfBunch, get_session_path, is_uuid_string,
                     load_file_content, remove_uuid_file)
 from ibllib.io import hashfile
-from ibllib.io.one import OneAbstract
 from ibllib.misc import pprint
 from oneibl.dataclass import SessionDataInfo
+from brainbox.io import parquet
+from brainbox.core import ismember, ismember2d
 
 _logger = logging.getLogger('ibllib')
 
@@ -90,28 +94,160 @@ SEARCH_TERMS = {  # keynames are possible input arguments and values are actual 
 }
 
 
-class ONE(OneAbstract):
-    def __init__(self, username=None, password=None, base_url=None, silent=False, printout=True):
+def _ses2pandas(ses, dtypes=None):
+    """
+    :param ses: session dictionary from rest endpoint
+    :param dtypes: list of dataset types
+    :return:
+    """
+    # selection: get relevant dtypes only if there is an url associated
+    rec = list(filter(lambda x: x['url'], ses['data_dataset_session_related']))
+    if dtypes == ['__all__'] or dtypes == '__all__':
+        dtypes = None
+    if dtypes is not None:
+        rec = list(filter(lambda x: x['dataset_type'] in dtypes, rec))
+    include = ['id', 'hash', 'dataset_type', 'name', 'file_size', 'collection']
+    uuid_fields = ['id', 'eid']
+    join = {'subject': ses['subject'], 'lab': ses['lab'], 'eid': ses['url'][-36:],
+            'start_time': np.datetime64(ses['start_time']), 'number': ses['number'],
+            'task_protocol': ses['task_protocol']}
+    col = parquet.rec2col(rec, include=include, uuid_fields=uuid_fields, join=join,
+                          types={'file_size': np.double}).to_df()
+    return col
+
+
+class OneAbstract(abc.ABC):
+
+    def __init__(self, username=None, password=None, base_url=None, cache_dir=None, silent=None):
         # get parameters override if inputs provided
         self._par = oneibl.params.get(silent=silent)
         self._par = self._par.set('ALYX_LOGIN', username or self._par.ALYX_LOGIN)
         self._par = self._par.set('ALYX_URL', base_url or self._par.ALYX_URL)
         self._par = self._par.set('ALYX_PWD', password or self._par.ALYX_PWD)
+        self._par = self._par.set('CACHE_DIR', cache_dir or self._par.CACHE_DIR)
+        # init the cache file
+        self._cache_file = Path(self._par.CACHE_DIR).joinpath('.one_cache.parquet')
+        if self._cache_file.exists():
+            # we need to keep this part fast enough for transient objects
+            self._cache = parquet.load(self._cache_file)
+        else:
+            self._cache = pd.DataFrame()
 
+    def _load(self, eid, dataset_types=None, dclass_output=False, download_only=False,
+              offline=False, **kwargs):
+        """
+        From a Session ID and dataset types, queries Alyx database, downloads the data
+        from Globus, and loads into numpy array. Single session only
+        """
+        if is_uuid_string(eid):
+            eid = '/sessions/' + eid
+        eid_str = eid[-36:]
+        # if no dataset_type is provided:
+        # a) force the output to be a dictionary that provides context to the data
+        # b) download all types that have a data url specified whithin the alf folder
+        dataset_types = [dataset_types] if isinstance(dataset_types, str) else dataset_types
+        if not dataset_types or dataset_types == ['__all__']:
+            dclass_output = True
+        if offline:
+            dc = self._make_dataclass_offline(eid_str, dataset_types, **kwargs)
+        else:
+            dc = self._make_dataclass(eid_str, dataset_types, **kwargs)
+        # load the files content in variables if requested
+        if not download_only:
+            for ind, fil in enumerate(dc.local_path):
+                dc.data[ind] = load_file_content(fil)
+        # parse output arguments
+        if dclass_output:
+            return dc
+        # if required, parse the output as a list that matches dataset_types requested
+        list_out = []
+        for dt in dataset_types:
+            if dt not in dc.dataset_type:
+                _logger.warning('dataset ' + dt + ' not found for session: ' + eid_str)
+                list_out.append(None)
+                continue
+            for i, x, in enumerate(dc.dataset_type):
+                if dt == x:
+                    if dc.data[i] is not None:
+                        list_out.append(dc.data[i])
+                    else:
+                        list_out.append(dc.local_path[i])
+        return list_out
+
+    def _get_cache_dir(self, cache_dir):
+        if not cache_dir:
+            cache_dir = self._par.CACHE_DIR
+        # if empty in parameter file, do not allow and set default
+        if not cache_dir:
+            cache_dir = str(PurePath(Path.home(), "Downloads", "FlatIron"))
+        return cache_dir
+
+    def _make_dataclass_offline(self, eid, dataset_types=None, cache_dir=None, **kwargs):
+        if self._cache.size == 0:
+            return SessionDataInfo()
+        # select the session
+        npeid = parquet.str2np(eid)[0]
+        df = self._cache[self._cache['eid_0'] == npeid[0]]
+        df = df[df['eid_1'] == npeid[1]]
+        # select datasets
+        df = df[ismember(df['dataset_type'], dataset_types)[0]]
+        return SessionDataInfo.from_pandas(df, self._get_cache_dir(cache_dir))
+
+    @abc.abstractmethod
+    def _make_dataclass(self, eid, dataset_types=None, cache_dir=None, **kwargs):
+        pass
+
+    @abc.abstractmethod
+    def load(self, **kwargs):
+        pass
+
+    @abc.abstractmethod
+    def list(self, **kwargs):
+        pass
+
+    @abc.abstractmethod
+    def search(self, **kwargs):
+        pass
+
+
+def ONE(offline=False, **kwargs):
+    if offline:
+        return OneOffline(**kwargs)
+    else:
+        return OneAlyx(**kwargs)
+
+
+class OneOffline(OneAbstract):
+
+    def _make_dataclass(self, *args, **kwargs):
+        return self._make_dataclass_offline(*args, **kwargs)
+
+    def load(self, eid, **kwargs):
+        return self._load(eid, **kwargs)
+
+    def list(self, **kwargs):
+        pass
+
+    def search(self, **kwargs):
+        pass
+
+
+class OneAlyx(OneAbstract):
+    def __init__(self, **kwargs):
+        # get parameters override if inputs provided
+        super(OneAlyx, self).__init__(**kwargs)
         try:
             self._alyxClient = wc.AlyxClient(username=self._par.ALYX_LOGIN,
                                              password=self._par.ALYX_PWD,
                                              base_url=self._par.ALYX_URL)
+            # Display output when instantiating ONE
+            print(f"Connected to {self._par.ALYX_URL} as {self._par.ALYX_LOGIN}", )
         except requests.exceptions.ConnectionError:
             raise ConnectionError(
                 f"Can't connect to {self._par.ALYX_URL}.\n" +
                 "IP addresses are filtered on IBL database servers. \n" +
                 "Are you connecting from an IBL participating institution ?"
             )
-        # Init connection to Globus if needed
-        # Display output when instantiating ONE
-        if printout:
-            print(f"Connected to {self._par.ALYX_URL} as {self._par.ALYX_LOGIN}",)
 
     @property
     def alyx(self):
@@ -292,31 +428,17 @@ class ONE(OneAbstract):
                     out.append(self._load(e, **kwargs)[0])
             return out
 
-    def _load(self, eid, dataset_types=None, dclass_output=False, dry_run=False, cache_dir=None,
-              download_only=False, clobber=False, offline=False, keep_uuid=False):
-        """
-        From a Session ID and dataset types, queries Alyx database, downloads the data
-        from Globus, and loads into numpy array. Single session only
-        """
+    def _make_dataclass(self, eid, dataset_types=None, cache_dir=None, dry_run=False,
+                        clobber=False, offline=False, keep_uuid=False):
         # if the input as an UUID, add the beginning of URL to it
         cache_dir = self._get_cache_dir(cache_dir)
-        if is_uuid_string(eid):
-            eid = '/sessions/' + eid
-        eid_str = eid[-36:]
         # get session json information as a dictionary from the alyx API
         try:
-            ses = self.alyx.get('/sessions/' + eid_str)
+            ses = self.alyx.rest('sessions', 'read', id=eid)
         except requests.HTTPError:
-            raise requests.HTTPError('Session ' + eid_str + ' does not exist')
-        # ses = ses[0]
-        # if no dataset_type is provided:
-        # a) force the output to be a dictionary that provides context to the data
-        # b) download all types that have a data url specified whithin the alf folder
-        dataset_types = [dataset_types] if isinstance(dataset_types, str) else dataset_types
-        if not dataset_types or dataset_types == ['__all__']:
-            dclass_output = True
-        # this performs the filtering
-        dc = SessionDataInfo.from_session_details(ses, dataset_types=dataset_types, eid=eid_str)
+            raise requests.HTTPError('Session ' + eid + ' does not exist')
+        # filter by dataset types
+        dc = SessionDataInfo.from_session_details(ses, dataset_types=dataset_types, eid=eid)
         # loop over each dataset and download if necessary
         with concurrent.futures.ThreadPoolExecutor(max_workers=NTHREADS) as executor:
             futures = []
@@ -333,27 +455,9 @@ class ONE(OneAbstract):
                 if future is None:
                     continue
                 dc.local_path[ind] = future.result()
-        # load the files content in variables if requested
-        if not download_only:
-            for ind, fil in enumerate(dc.local_path):
-                dc.data[ind] = load_file_content(fil)
-        # parse output arguments
-        if dclass_output:
-            return dc
-        # if required, parse the output as a list that matches dataset_types requested
-        list_out = []
-        for dt in dataset_types:
-            if dt not in dc.dataset_type:
-                _logger.warning('dataset ' + dt + ' not found for session: ' + eid_str)
-                list_out.append(None)
-                continue
-            for i, x, in enumerate(dc.dataset_type):
-                if dt == x:
-                    if dc.data[i] is not None:
-                        list_out.append(dc.data[i])
-                    else:
-                        list_out.append(dc.local_path[i])
-        return list_out
+        # filter by daataset types and update the cache
+        self._update_cache(ses, dataset_types=dataset_types)
+        return dc
 
     def _ls(self, table=None, verbose=False):
         """
@@ -488,14 +592,6 @@ class ONE(OneAbstract):
         else:
             return eids
 
-    def _get_cache_dir(self, cache_dir):
-        if not cache_dir:
-            cache_dir = self._par.CACHE_DIR
-        # if empty in parameter file, do not allow and set default
-        if not cache_dir:
-            cache_dir = str(PurePath(Path.home(), "Downloads", "FlatIron"))
-        return cache_dir
-
     def download_datasets(self, dsets, **kwargs):
         """
         Download several datsets through a list of alyx REST dictionaries
@@ -528,6 +624,18 @@ class ONE(OneAbstract):
         target_dir = Path(self._get_cache_dir(cache_dir), relpath)
         return self._download_file(url=url, target_dir=target_dir, **kwargs)
 
+    def _tag_mismatched_file_record(self, url):
+        fr = self.alyx.rest('files', 'list', django=f"dataset,{Path(url).name.split('.')[-2]},"
+                                                    f"data_repository__globus_is_personal,False")
+        if len(fr) > 0:
+            json_field = fr[0]['json']
+            if json_field is None:
+                json_field = {'mismatch_hash': True}
+            else:
+                json_field.update({'mismatch_hash': True})
+            self.alyx.rest('files', 'partial_update', id=fr[0]['url'][-36:],
+                           data={'json': json_field})
+
     def _download_file(self, url, target_dir, clobber=False, offline=False, keep_uuid=False,
                        file_size=None, hash=None):
         """
@@ -545,23 +653,27 @@ class ONE(OneAbstract):
         local_path = str(target_dir) + os.sep + os.path.basename(url)
         if not keep_uuid:
             local_path = remove_uuid_file(local_path, dry=True)
-        if Path(local_path).exists():
-            # overwrites the file if the expected filesize is different from the cached filesize
-            if file_size and Path(local_path).stat().st_size != file_size:
+        if Path(local_path).exists() and not offline:
+            # the local file hash doesn't match the dataset table cached hash
+            hash_mismatch = hash and hashfile.md5(Path(local_path)) != hash
+            file_size_mismatch = file_size and Path(local_path).stat().st_size != file_size
+            if hash_mismatch or file_size_mismatch:
                 clobber = True
-            # overwrites the file if the expected hash is different from the cached hash
-            if hash and hashfile.md5(Path(local_path)) != hash:
-                clobber = True
+                _logger.warning(f" local md5 or size mismatch, re-downloading {local_path}")
         # if there is no cached file, download
         else:
             clobber = True
         if clobber:
-            local_path = wc.http_download_file(url,
-                                               username=self._par.HTTP_DATA_SERVER_LOGIN,
-                                               password=self._par.HTTP_DATA_SERVER_PWD,
-                                               cache_dir=str(target_dir),
-                                               clobber=clobber,
-                                               offline=offline)
+            local_path, md5 = wc.http_download_file(
+                url, username=self._par.HTTP_DATA_SERVER_LOGIN,
+                password=self._par.HTTP_DATA_SERVER_PWD, cache_dir=str(target_dir),
+                clobber=clobber, offline=offline, return_md5=True)
+            # post download, if there is a mismatch between Alyx and the newly downloaded file size
+            # or hash flag the offending file record in Alyx for database maintenance
+            hash_mismatch = hash and md5 != hash
+            file_size_mismatch = file_size and Path(local_path).stat().st_size != file_size
+            if hash_mismatch or file_size_mismatch:
+                self._tag_mismatched_file_record(url)
         if keep_uuid:
             return local_path
         else:
@@ -594,31 +706,50 @@ class ONE(OneAbstract):
         """
         oneibl.params.setup()
 
-    def path_from_eid(self, eid: str, grep_str=None) -> Path:
+    def path_from_eid(self, eid: str, use_cache=True) -> Path:
+        """
+        From an experiment id or a list of experiment ids, gets the local cache path
+        :param eid: eid (UUID) or list of UUIDs
+        :param use_cache: if set to False, will force database connection
+        :return: eid or list of eids
+        """
         # If eid is a list of eIDs recurse through list and return the results
         if isinstance(eid, list):
             path_list = []
             for p in eid:
-                path_list.append(self.path_from_eid(p, grep_str=grep_str))
+                path_list.append(self.path_from_eid(p))
             return path_list
         # If not valid return None
         if not is_uuid_string(eid):
             print(eid, " is not a valid eID/UUID string")
             return
-        # Load data, if no data present on disk return None
-        data = self._load(eid, download_only=True, offline=True)
-        if not data.local_path:
+
+        # first try avoid hitting the database
+        if self._cache.size > 0 and use_cache:
+            ic = parquet.find_first_2d(
+                self._cache[['eid_0', 'eid_1']].to_numpy(), parquet.str2np(eid))
+            if ic is not None:
+                ses = self._cache.iloc[ic]
+                return Path(self._par.CACHE_DIR).joinpath(
+                    ses['lab'], 'Subjects', ses['subject'], ses['start_time'].isoformat()[:10],
+                    str(ses['number']).zfill(3))
+
+        # if it wasn't successful, query Alyx
+        ses = self.alyx.rest('sessions', 'list', django=f'pk,{eid}')
+        if len(ses) == 0:
             return None
-        # If user defined a grep list of specific files return paths to files
-        if grep_str is not None:
-            files = [x for x in data.local_path if grep_str in str(x)]
-            return files
-        # If none of the above happen return the session path of the first file you find
-        session_path = get_session_path(data.local_path[0])
+        else:
+            return Path(self._par.CACHE_DIR).joinpath(
+                ses[0]['lab'], 'Subjects', ses[0]['subject'], ses[0]['start_time'][:10],
+                str(ses[0]['number']).zfill(3))
 
-        return session_path
-
-    def eid_from_path(self, path_obj):
+    def eid_from_path(self, path_obj, use_cache=True):
+        """
+        From a local path, gets the experiment id
+        :param path_obj: local path or list of local paths
+        :param use_cache: if set to False, will force database connection
+        :return: eid or list of eids
+        """
         # If path_obj is a list recurse through it and return a list
         if isinstance(path_obj, list):
             path_obj = [Path(x) for x in path_obj]
@@ -632,10 +763,22 @@ class ONE(OneAbstract):
         # if path does not have a date and a number return None
         if session_path is None:
             return None
-        # search for subj, date, number XXX: hits the DB
+
+        # try the cached info to possibly avoid hitting database
+        if self._cache.size > 0 and use_cache:
+            ind = ((self._cache['subject'] == session_path.parts[-3]) &
+                   (self._cache['start_time'].apply(
+                       lambda x: x.isoformat()[:10] == session_path.parts[-2])) &
+                   (self._cache['number']) == int(session_path.parts[-1]))
+            ind = np.where(ind.to_numpy())[0]
+            if ind.size > 0:
+                return parquet.np2str(self._cache[['eid_0', 'eid_1']].iloc[ind[0]])
+
+        # if not search for subj, date, number XXX: hits the DB
         uuid = self.search(subjects=session_path.parts[-3],
                            date_range=session_path.parts[-2],
                            number=session_path.parts[-1])
+
         # Return the uuid if any
         return uuid[0] if uuid else None
 
@@ -663,6 +806,49 @@ class ONE(OneAbstract):
         out = {k: v for k, v in dets.items() if k in det_fields}
         out.update({'local_path': self.path_from_eid(eid)})
         return out
+
+    def _update_cache(self, ses, dataset_types):
+        """
+        :param ses: session details dictionary as per Alyx response
+        :param dataset_types:
+        :return: is_updated (bool): if the cache was updated or not
+        """
+        save = False
+        pqt_dsets = _ses2pandas(ses, dtypes=dataset_types)
+        # if the dataframe is empty, return
+        if pqt_dsets.size == 0:
+            return
+        # if the cache is empty create the cache variable
+        elif self._cache.size == 0:
+            self._cache = pqt_dsets
+            save = True
+        # the cache is not empty and there are datasets in the query
+        else:
+            isin, icache = ismember2d(pqt_dsets[['id_0', 'id_1']].to_numpy(),
+                                      self._cache[['id_0', 'id_1']].to_numpy())
+            # check if the hash / filesize fields have changed on patching
+            heq = (self._cache['hash'].iloc[icache].to_numpy() ==
+                   pqt_dsets['hash'].iloc[isin].to_numpy())
+            feq = np.isclose(self._cache['file_size'].iloc[icache].to_numpy(),
+                             pqt_dsets['file_size'].iloc[isin].to_numpy(),
+                             rtol=0, atol=0, equal_nan=True)
+            eq = np.logical_and(heq, feq)
+            # update new hash / filesizes
+            if not np.all(eq):
+                self._cache.iloc[icache].loc[:, ['file_size', 'hash']] =\
+                    pqt_dsets.iloc[np.where(isin)[0]].loc[:, ['file_size', 'hash']]
+                save = True
+            # append datasets that haven't been found
+            if not np.all(isin):
+                self._cache = self._cache.append(pqt_dsets.iloc[np.where(~isin)[0]])
+                self._cache = self._cache.reindex()
+                save = True
+        if save:
+            # before saving makes sure pandas did not cast uuids in float
+            typs = [t for t, k in zip(self._cache.dtypes, self._cache.keys()) if 'id_' in k]
+            assert (all(map(lambda t: t == np.int64, typs)))
+            # if this gets too big, look into saving only when destroying the ONE object
+            parquet.save(self._cache_file, self._cache)
 
 
 def _validate_date_range(date_range):