+++ release_notes.md
@@ -1,3 +1,11 @@
+### Release Notes 1.5.5
+-   ONE offline mode and cache dataset table to speed up reloading of large datasets (Olivier)
+-   ALF io naming conventions on loading objects (Miles)
+-   KS2 Matlab ephys pipeline tasks (Olivier)
+-   Support for running QC on biased and training sessions (Nico)
+-   metrics_df and passed_df properties in BpodQC obj for qcplots (Nico)
+-   Added missing unittest to stim_move_before_goCue metric (Nico)
+
 ### Release Notes 1.5.4 - 29/07/2020 hotfix
 -   ibllib.pipes.training_preprocessing.TrainingAudio